video-generator/requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
googletrans==3.0.0
gTTS==2.1.1
gTTS-token==1.1.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.15
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
moviepy==1.0.3
numpy==1.19.0
Pillow==7.2.0
proglog==0.1.9
requests==2.24.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
soupsieve==2.0.1
tqdm==4.47.0
urllib3==1.25.9
